refactor(src): split worktree/mod.rs by responsibility
